+++ src/Generator/Taxonomy.php
@@ -21,142 +21,62 @@ use Cecil\Exception\Exception;
  */
 class Taxonomy extends AbstractGenerator implements GeneratorInterface
 {
-    /* @var VocabulariesCollection */
-    protected $vocabulariesCollection;
-
     /**
      * {@inheritdoc}
      */
     public function generate(): void
     {
         if ($this->config->get('site.taxonomies')) {
-            $this->createVocabulariesCollection();
-            $this->collectTermsFromPages();
-            $this->generateTaxonomiesPages();
-        }
-    }
-
-    /**
-     * Create a collection from the vocabularies configuration.
-     */
-    protected function createVocabulariesCollection()
-    {
-        // create an empty a vocabularies collection
-        $this->vocabulariesCollection = new VocabulariesCollection('taxonomies');
-        /*
-         * Adds each vocabulary to the collection.
-         * ie:
-         *   taxonomies:
-         *     - tags: tag
-         *     - categories: category
-         */
-        foreach (array_keys($this->config->get('site.taxonomies')) as $vocabulary) {
-            /*
-             * Disabled vocabulary?
-             * ie:
-             *   taxonomies:
-             *     tags: disabled
-             */
-            if ($this->config->get("site.taxonomies.$vocabulary") == 'disabled') {
-                continue;
-            }
-
-            $this->vocabulariesCollection->add(new Vocabulary($vocabulary));
-        }
-    }
-
-    /**
-     * Collects vocabularies/terms from pages frontmatter.
-     */
-    protected function collectTermsFromPages()
-    {
-        /* @var $page Page */
-        $pages = $this->pagesCollection->sortByDate();
-        foreach ($pages as $page) {
-            // ie: tags
-            foreach ($this->vocabulariesCollection as $vocabulary) {
+            /* @var $vocabulary Vocabulary */
+            foreach ($this->builder->getTaxonomies() as $position => $vocabulary) {
                 $plural = $vocabulary->getId();
-                /*
-                 * ie:
-                 *   tags: Tag 1, Tag 2
-                 */
-                if ($page->hasVariable($plural)) {
-                    // converts a string list to an array
-                    if (!is_array($page->getVariable($plural))) {
-                        $page->setVariable($plural, [$page->getVariable($plural)]);
+                $singular = $this->config->get("site.taxonomies.$plural");
+                if (count($vocabulary) > 0) {
+                    /*
+                    * Creates $plural/$term pages (list of pages)
+                    * ie: /tags/tag-1/
+                    */
+                    /* @var $pages PagesCollection */
+                    foreach ($vocabulary as $position => $term) {
+                        $pageId = $path = Page::slugify(sprintf('%s/%s', $plural, $term->getId()));
+                        $pages = $term->sortByDate();
+                        $date = $pages->first()->getVariable('date');
+                        $page = (new Page($pageId))
+                            ->setVariable('title', $term->getName());
+                        if ($this->pagesCollection->has($pageId)) {
+                            $page = clone $this->pagesCollection->get($pageId);
+                        }
+                        $page
+                            ->setType(Type::TERM)
+                            ->setPath($path)
+                            ->setVariable('date', $date)
+                            ->setVariable('term', $term->getId())
+                            ->setVariable('plural', $plural)
+                            ->setVariable('singular', $singular)
+                            ->setVariable('pages', $pages)
+                            ->setVariable('pagination', ['pages' => $pages]);
+                        $this->generatedPages->add($page);
                     }
-                    // adds each term to the vocabulary collection...
-                    foreach ($page->getVariable($plural) as $termName) {
-                        $termId = Page::slugify($termName);
-                        $term = (new Term($termId))->setName($termName);
-                        $this->vocabulariesCollection
-                            ->get($plural)
-                            ->add($term);
-                        // ... and adds page to the term collection
-                        $this->vocabulariesCollection
-                            ->get($plural)
-                            ->get($termId)
-                            ->add($page);
-                    }
-                }
-            }
-        }
-    }
-
-    /**
-     * Generate taxonomies pages.
-     */
-    protected function generateTaxonomiesPages()
-    {
-        /* @var $vocabulary Vocabulary */
-        foreach ($this->vocabulariesCollection as $position => $vocabulary) {
-            $plural = $vocabulary->getId();
-            $singular = $this->config->get("site.taxonomies.$plural");
-            if (count($vocabulary) > 0) {
-                /*
-                 * Creates $plural/$term pages (list of pages)
-                 * ie: /tags/tag-1/
-                 */
-                /* @var $pages PagesCollection */
-                foreach ($vocabulary as $position => $term) {
-                    $pageId = $path = Page::slugify(sprintf('%s/%s', $plural, $term->getId()));
-                    $pages = $term->sortByDate();
-                    $date = $pages->first()->getVariable('date');
+                    /*
+                    * Creates $plural pages (list of terms)
+                    * ex: /tags/
+                    */
+                    $pageId = $path = Page::slugify($plural);
                     $page = (new Page($pageId))
-                        ->setVariable('title', $term->getName());
-                    if ($this->pagesCollection->has($pageId)) {
-                        $page = clone $this->pagesCollection->get($pageId);
-                    }
-                    $page
-                        ->setType(Type::TERM)
+                        ->setType(Type::VOCABULARY)
                         ->setPath($path)
+                        ->setVariable('title', ucfirst($plural))
                         ->setVariable('date', $date)
-                        ->setVariable('term', $term->getId())
                         ->setVariable('plural', $plural)
                         ->setVariable('singular', $singular)
-                        ->setVariable('pages', $pages)
-                        ->setVariable('pagination', ['pages' => $pages]);
-                    $this->generatedPages->add($page);
-                }
-                /*
-                 * Creates $plural pages (list of terms)
-                 * ex: /tags/
-                 */
-                $pageId = $path = Page::slugify($plural);
-                $page = (new Page($pageId))
-                    ->setType(Type::VOCABULARY)
-                    ->setPath($path)
-                    ->setVariable('title', ucfirst($plural))
-                    ->setVariable('date', $date)
-                    ->setVariable('plural', $plural)
-                    ->setVariable('singular', $singular)
-                    ->setVariable('terms', $vocabulary);
-                // add page only if a template exist
-                try {
-                    $this->generatedPages->add($page);
-                } catch (Exception $e) {
-                    printf("%s\n", $e->getMessage());
-                    unset($page); // do not add page
+                        ->setVariable('terms', $vocabulary);
+                    // add page only if a template exist
+                    try {
+                        $this->generatedPages->add($page);
+                    } catch (Exception $e) {
+                        printf("%s\n", $e->getMessage());
+                        unset($page); // do not add page
+                    }
                 }
             }
         }
